Add GUIControl class.
Make Talents not directly interfacing with QML.

This is a step towards making Talents internal per Character.
The end goal is to be able to have multiple classes talents available
via a quick class switch bar. This is currently not easily
implemented when Talents is a shared instance among all classes and
directly facing QML.