Make GUIControl initialize all classes in use.
Make GUIControl destroy these objects. This is mostly for good practice:
if GUIControl ends, the application is finished anyway.

Add class color to C++ portion as well (previously only in QML). The
reasoning here is that in the future it is likely that we want this to
dump .html files with the result and then we would like to include class
color somewhere.

Make class properties name and color available.

Remove hardcoded class name and color from TalentCalculator.qml.