Use enable/disable for Unbridled Wrath.
Enable/disabled Unbridled Wrath when spending/refunding talent points.
Previously Unbridled Wrath was always in the proc list, which meant it
was always evaluated even if it could not proc.

This also means there was always a statistics entry for it. This was
hidden with the "ignore procs with 0 successful procs" filter currently
in place in ClassStatistics::get_proc_table().