Fix issues raised by static code analysis.
This continues the work in e7226656a991cda55c324d42894f3fb9eb3c80d9 by
removing further places where we copy structures.

Bonus: refactor the usage of strings to instead use int. This is
achieved by creating a TalentTier class and removing the usage of QMap
with a QString as key.