Refactor constructor arguments.
Constructors with changed arguments:
- Character
- CharacterStats
- CombatRoll
- Equipment
- EquipmentDb
- TestBuff
- TestProc
- TestSpell
- TestWarrior
- ..and for classes inheriting any of these (which are many)

Character has assumed ownership of many of its arguments and creates
them itself. This reduces a lot of boiler-plate code.

Equipment and EquipmentDb has had its dependencies reversed, Equipment
now requires an EquipmentDb as argument. This allows instantiating a
single EquipmentDb (which reads all the equipment files it can find,
which is slow) and passing it around to all created Equipment instances.

With this reversed dependency, all tests are modified to pass around the
same instance of EquipmentDb. This is a significant performance
improvement.

CombatRoll takes Character instead of Target as argument. With this
change we can remove the set_character() function in Target (since
Target is created per Character, instead of previously shared between
multiple Characters).