Remove set_character from Equipment.
Before commit 8500f88605a15a8c0f606954c4a7077180a39151 Equipment needed
to be able to handle many Character instances via set_character. With
the mentioned commit we instead have a 1:1 mapping of Character ->
Equipment (Equipment is now owned by Character). This means
set_character no longer makes any sense as a function.

set_character used to clean up item state when it was invoked, but with
its removal there is still a need to clean up item state even if the
Character does not change.

This is because in the mentioned commit EquipmentDb is much more
persistent between simulations, so it stores Items that have Character
state in them.

The base reason is due to how spells (procs in this case) work. Spells are
highly coupled with pchar (which in itself is not that problematic) and
in particular the derived class instances like Warrior (which is more
problematic). The reason for this is, among others, the access to the
derived ClassStatistics (WarriorStatistics*).

We cannot destroy Equipment and clean item state via Equipment
destructor, because spells try to access ClassStatistics but it has
already been destroyed (it belongs to e.g. Warrior, while Equipment
belong to Character). Therefore we need to manually clean the item
state from Equipment before destroying Equipment (via destruction of
pchar).