Let Equipment handle copies from Db.
Previously Equipment handled the same object as stored in the database.
With changes to Character - Equipment mapping in
8500f88605a15a8c0f606954c4a7077180a39151 items stored in the EquipmentDb
were previously touched by all Character instances.

With this change each Equipment instance can modify its Item (and
mainly the procs connected to it) at will. The cost is that Equipment
must make sure to free any objects it has retrieved from the
EquipmentDb.