Fix a max talent bug in Buff.
When going via both interface function enable_buff() and increase_rank()
the rank was increased twice, since enable_buff() calls increase_rank().
This caused 5/5 Flurry to get array index out of bounds fault.

Do not use talent rank for checking if enabled, use regular enable
instead. Most buffs do not have talent ranks.

Add assertion that max talent is not crossed.

Also, update curr_points before applying/removing rank effect. This
leads to slightly cleaner code, and it makes sense for the
implementations of rank effect to not need to have the knowledge that
curr_points is not actually the current rank, but is either previous or
the next rank.