Add Slam.
Add CastComplete event in order to delay the effect of spell damage.
Note that currently this is a duplicate of ResourceGain. CastComplete
should use another interface function than perform_periodic() in the
future.

Add reset_swingtimer() to attack spells in order to support slam
behavior of resetting swing timers upon cast completion. Also add to
attack spells to ignore events if character not attacking. This is
required to model the behavior of Slam suspending melee attacks while
casting.

Improve assertion prints in test suite. Also add possiblity to act on
given event (default is to not act).