Unify talent handling.
Add TalentRequirer class. Anything that can be modified by talents
inherits this class on top of any other base class. Talent modifications
are now implemented in the spell/buff that is modified. It is (1)
separate from the Talent class, and (2) close to e.g. damage calculation
code, which is an added bonus.

Remove the current interface functions that not only had different names
but only had different implementations (in Spell.h and Buff.h).

With this change talents can now only be modified by actually
instantiating and invoking the talent class itself, since only it has
access to the talent modifying functions. This clears up the interface
for how to modify something via talents (e.g. force us to go via
talent, not something in between).