Remove unnecessary assert.
This enables swapping rotations.

Link spells is only needed once in the current setup, the second time
around the spells are linked and this assert fails. This may change
if/when we look at whether the spells is enabled before linking it.

Comment rotation.xml files that contain unimplemented spell and
unsupported variable assignment. With the applied changes the files may
be copied to the build folder and work immediately.