Support common statistics object across threads.
Add NumberCruncher class. This class is an object that is shared between
threads. It holds the ultimate result for a simulation under a given
option. Because it supports many threads, under the option key is a
vector of results (ClassStatistics instances).

Move ownership of Statistics classes from Spell/Buff/Proc to
ClassStatistics. This is such that persistency of statistics is no
longer per spell, but rather per simulation. For each new simulation a
new statistics instance is available.

Remove reset when removing active buffs. This is problematic with the
change above when clearing talent trees. We should not need to reset
things when we deactivate them, if anything it should be the other way
around: make sure things are initialized properly when preparing for
combat.

Let ClassStatistics store vectors (from threads) of results per option.
These will be parsed in upcoming commits for getting the value of stat
scaling.

Add NoScale option to get a baseline of values to compare to.

Fix Sword Spec not being deleted.

Remove unused functions of type get_statistics() from Buff/Proc.