Add layout direction and z-value.
Layout direction is to be able to show bracer enchants upwards rather
than downwards. Z-value is to fix the shoulder-model is shown under back
enchants, which makes it difficult to click on certain shoulder
enchants. Just switching the order of QML elements would not have worked
as a real solution, as this problem will be present in future places as
well (notably weapon enchants).