Resolve circular dependency during destruction.
There is a circular dependency involving Equipment, CharacterStats,
ActiveBuffs, GeneralBuffs, and ExternalBuff.

This can be seen when using Rallying Cry of the Dragonslayer.

Previously CharacterStats destroyed Equipment during its destruction.
Equipment contains items that may have procs on them, so
ActiveBuffs must not be destroyed when Equipment is destroyed.

In the other direction, ActiveBuffs destroys GeneralBuffs, which
contains ExternalBuff, and these often (always?) expect CharacterStats
to not be destroyed.

The solution is to unequip all gear and cancel all buffs/procs before
deleting objects.