Make Rotation non-abstract.
Remove WarriorRotation by adding get_spell/buff_by_name() functions.
This can be used to get spells provided by items as well.

Also remove the concept of atomic rotations, as in either parsed
completely or not at all. This is because we go via the current active
spells, which means plenty of rotations are not able to be parsed most
of the time (i.e. cannot switch to a Fury rotation if BT not talented).
Instead we relink the spells when such a condition that provides a spell
changes (equipment change, etc).

There is a potential for running "partial" rotations, in that they may
be difficult to debug. This will be much easier once the rotation
statistics is in place (with its executor counts, etc).