Add Jom Gabbar.
Add PeriodicRefreshBuff event that calls a new refresh_buff(). The
difference between refresh_buff() and apply_buff() is that the former
does not extend the duration of the buff, while still invoking the
refresh buff effect function (that in Jom Gabbar's case increases attack
power).

Remove some strange if-assertions that had been placed for breakpoint
purposes.