Add Badge of the Swarmguard.
It has an interesting effect in that when activated it has zero
immediate effect. It only enables a proc to actually start stacking a
buff. Once 30 seconds have passed from the initial activation all stacks
are removed.

This is solved by adding a NoEffect aura that is linked with a proc and
a buff. The link to the proc is used to make the proc only available
when the buff is active. The link to the armor penetration buff is to
make sure it is cancelled once the initial 30 seconds have passed,
regardless of the remaining duration of the armor penetration buff.