Refactor resource handling.
Spells now must define a resource type in order to know which resource
they should compare to.

Let Character (for now, may refactor to separate class) handle all
resources rather than letting class implementations themselves handle
it. This is in order to provide a generic interface for spells.

Improve ConditionResource such that it actually takes into account which
resource it compares to. This would not have worked very well for
druids.