Refactor resource handling.
Every playable class now has its resource as a member.

Resources are now in separate classes instead of part of Character and
subclasses to Character.

Add Resource class to define the basic interface of the subclasses.

Add Rage and allow Druid and Warrior to access its internal members.

Add Energy and allow Druid and Rogue to access its internal members.

Add Mana and allow the rest of the classes (+ Druid) to access its
internal members.