Add TestTalentTree to define tree tests.
Use TestTalentTree as base in TestFury and TestArms.

Also add helper functions that goes via Talents-class instead of
directly via TalentTree. This means a slightly higher level of testing
is done, as it implicitly tests some of the Talents-class as well.

It allows testing that talents in each TalentTree implementation
correctly resets between talent setups, which is not possible otherwise.