Remove default value for crit mod.
Crit mod defaulted to 0.0 (from the time when this *was* a pure crit mod)
for melee hit result and base mainhand / offhand did not send in pchar
crit rate, which meant that the default value of 0.0 was used, leading
to zero crits for Rogue from MH / OH attacks.

Warriors are unaffected by this bug because they use
MainhandAttackWarrior and OffhandAttackWarrior, which are inherited
classes that DO send in pchar crit rate.