Make Instant Poison deal non-zero damage.
Instant Poisons had a few issues as described below.

Proc effects could not occur since poisons were not applied correctly.
When overriding spell_effect as a Proc derivative (which was done in
order to apply_buff()) the proc check no longer runs, as it is
overriden.

This prompted the fix in 0b697077282be1ebb3ed70d2b5a77b28a388ad44 which
separates the program flow between performing a pre-combat action and an
in combat action. This is mostly a sign that Instant Poisons could be
implemented using 3 spells rather than 2 (one Spell, one Proc, one Buff)
but keeping the number of objects required to implement a single
damaging effect low is desirable.

Also, vile_poisons was not instantiated which lead to 0.0 modifier that
reduced all damage to zero even when proc check was fixed.

Also, Proc did not invoke its spell specific effect on set of pre combat
iterations.