Add Consecrated Sharpening Stones.
Fix bug that AP gains vs creatures were not included if they were not on
equipment (so e.g. provided via talents, or in this case a sharpening
stone). Note that there were no previous use-case for this scenario so
the bug was merely dormant until now, rather than something that
affected previous results.