Re-equip items on patch selection.
Previously items were only cleared if patch went below the current patch
of the item. This however did not work properly since the point of
comparison was the value for <patch> and did not take into consideration
if the item was still available in the patch, just with different data.

Example:
Equip Felstriker in 1.12. Change patch to 1.0.0. Then Felstriker is
cleared entirely, although it is available under the name Deathstriker.

With this commit the example changes to:
Equip Felstriker in 1.12. Change patch to 1.0.0. Then Felstriker is
named Deathstriker.

Another thing that did not work was that when selecting a later patch
that contained changes the item was not updated (or even cleared) to
reflect the change.

Example:
Equip Deathstriker in 1.0.0. Change patch to 1.3.0. Then the equipped
item is still named Deathstriker.

With this commit the example changes to:
Equip Deathstriker in 1.0.0. Change patch to 1.3.0. Then the equipped
item is named Felstriker.