Make buff models support multiple characters.
Previously BuffModel and DebuffModel had poor support for multiple
characters (as noted by the now removed comment in GUIControl
constructor). Faction / class change did not correctly update state of
models.

Also, let BuffModel / DebuffModel toggle buffs instead of GUIControl.
This inversion allows the buff models to immediately know when to update
the model data instead of being told to do so by GUIControl. Also, it
conceals this knowledge from GUIC and makes it more into solely an
interface (which is the intent of GUIC).