Change constructor signature.
Currently it is initialized with an uninitialized pointer, and it works
because the pointer is never used until it is replaced with a valid
pointer from the first set_character invocation.

By removing the unnecessary pointer it is more clear how the usage of
RotationModel is expected to be, and an invalid pointer is never
assigned.

Also, shuffle around some initializations in GUIControl constructor
such that all models are initialized next to each other.