Add Dense Sharpening Stone.
Because these sharpening stones are only available on edged weapons some
helper functions are useful to find out if current weapon for which we
are creating a model is indeed a sharp weapon. This leads to some
refactoring that allows moving code to base class CharacterEnchants.

Make sure to update the model whenever MH/OH changes such that the model
contains the sharpening stones if it should.

Save current character in EnchantModel to avoid sending set_character
with the current character.