Remove fixed flavor text height.
For long flavor texts (e.g. Taut Dragonhide Shoulderpads) a fixed text
height causes the text to overlap with elements above as well as leave
the box boundaries. This is because of a center vertical alignment.

Since height is calculated based on contents, simply removing the fixed
height fixes these issues.