Show set bonuses in item tooltip.
This requires a few changes in other places:

- EquipmentDb needs to track all item IDs in order to be able to quickly
respond to the question "for a given item id, what is its name?". The
alternative is a linear search which is unwieldy.
- Equipment needs to expose the SetBonusControl instance to GUI control.
Also remove unused function.
- Change some function signatures and data structures in set bonus
related classes in order to easily track which bonuses are active.
- Add set number information in set_bonuses.xml. This may be reworked to
use the <bonus value="num"> instead (it was the original intention) but
it needs more changes to make it work. This data needs to be propagated
with the tooltips, but is currently lost.
- Add helper function to SetBonusControl that GUIControl can use in
order to build up the tooltip data.
- Add the required text elements in EquipmentTooltip.qml.