Allow individual items to modify behavior.
Add ItemModificationRequirer class that provides an interface for
modifying behavior via item ID.

Add Blood Guard's Plate Gauntlets with a new <modifies> element that
specifies Hamstring should be notified when it is equipped/unequipped.

Inherit from the new ItemModificationRequirer class in Hamstring in
order to implement the modifications whenever a specific item id is
equipped, such that rage cost is reduced.