Do not break dynamic binding.
ttRect.visible uses highlight.visible to determine visibility. Many
subsequent tooltip elements uses ttRect.visible. When breaking that
dynamic binding on class change the tooltips stops appearing for the
talent slot locations that are not available for the initial class.

Example: Protection 4RR is not available for warriors. Assume warriors
are the initial class, then change to hunter. Hunters have Imp Feign
Death in Surv 4RR, but the tooltip does not appear unless Hunters are
the initial class, because the binding breaks when changing class.