Make Talent non-abstract.
Make two of the rank description functions static and remove the third.
The removal is based on it being an inferior version of one of the
remaining description functions.

Make Talent non-abstract in order to implement what has become generic
invocation patterns. Currently the only supported pattern is classes
inheriting Spell and TalentRequirer.

Move Arms talents that fit the restriction above to using the regular
Talent instead and remove the now unnecessary derived classes:

- ImprovedHeroicStrike
- ImprovedRend
- ImprovedOverpower
- DeepWoundsTalent
- MortalStrikeTalent

Also remove SweepingStrikes as it can be replaced by GenericTalent since
it has no effect.