Rework TestCharacter into TestObject.
Rename TestCharacter to TestObject.

Rename the test it mandates should be implemented to
test_values_after_initialization to be more specific what the test
purpose is.

Change the constructor signature to require an EquipmentDb pointer. This
reduces the expensive EquipmentDb setup to only be done once when
starting tests, and once again when starting the application.