Add time_since_auto_shot builtin variable.
Remove the setting of cooldown in the constructor. Let each set of
combat iterations decide what the cooldown should be as which item to
derive the cooldown from is known at that point.

Combined with resetting wheneter set of combat iterations are being
prepared allows for the proper usage of Spell::get_last_used() in
ConditionVariableBuiltin (since it uses cooldown, which previously was
not set correctly), except for the necessity of handling negative values
at start with std::max() since the spell has not yet been cast.

The auto attack spells are special in this regard in that they have a
very variable cooldown that is only known when actually starting to run
the simulation, which is why a reset is necessary to make sure the first
iteration runs as well as all the following iterations.

As a bonus: set the icon shown in statistics to be the equipped weapon
rather than a static hardcoded icon.