Add Pet class.
Add Cat class that constructs Pet class and instantiates Claw and
PetAutoAttack for use.

Add getter for resource_cost in Spell to enable a circumvention of
Spell::is_available() for Pet. This is because is_available uses pchar,
which for Pet spells is incorrect.

Make certain methods static in CharacterStats in order to use in Pet.

Add Focus as ResourceType in order to support Pet resource usage.