Add Bestial Wrath spell/talent.
Note that there is a circular dependency between HunterSpells and Pet
initialization: Pet wants to add spells to HunterSpells, while
HunterSpells wants the pet name. The circular dependency is currently
broken by instead of referring to a non-initialized Pet pointer in
HunterSpells we send the QString "Cat" as argument to BestialWrath
initialization in HunterSpells. This will have to be solved in the
future.