Fix Vaelastrasz ruleset not properly applied.
The application of the Vaelastrasz ruleset depended on the engine not
being reset, which happens as the last step before running the sim.

By adding a concept of pre-combat buffs like spells has the
EncounterStart can instead apply the buff rather than RulesetControl.
That puts the buff application within the proper domain of being inside
the event chain.