Track assets with Git LFS.
Each individual icon is tracked, which adds a hefty download time when
cloning (at the time of writing around ~20 minutes, and will keep
growing). An alternative is to add all images to an archive and upload
that instead, which requires all developers to update the archive and
extract before building. BUT, LFS also tracks the archive as modified at
all times, which is an annoyance so this option was discarded.