Reset characters and check Target cleanliness.
Check that all debuffs were properly cleared in Target to ensure that no
debuff leaves itself on the Target after reset.

Previously Target reset all buffs before Character, which lead to a
doublue reset of all buffs that were placed on Target, leading to a much
reduced reported uptime (the second reset reported an uptime of 0).