Refactor periodic spells.
In order to formalize and unify the handling of e.g. target debuffs
falling off before a dot tick and other such problems that has to be
considered for every debuff the periodic buffs are now a base class
SpellPeriodic.

Add a SpellCallback event for handling special cases. For now only Seal
of Command uses it and before formalizing it further the new
SpellCallback event only works with Seal of Command.

Improve resource and execution time statistics for periodic effects.
This is simpler now that it only has to be solved once.

Remove ResourceGain for the moment.