Refactor debuffs with charges handling.
The goal is to make the entry point for adding charge debuffs generic
and being able to handle the consumption events in an easier manner. It
must also be possible in the future to generically add ignoring charge
consumption even when the event occurs due to it being how some spells
work (e.g. Shadow Oil does not consume ISB charges).

Add checks that all the charge debuff vectors are empty when target has
been cleaned.