Support melee auto attacks in Cat Form.
Add MainhandAttackDruid in order to support attacks in Cat Form, Bear
Form, and Caster/Moonkin forms.

Add functions in Item and Weapon to enable/disable certain enchants that
should have no effect while in Druid forms.

Create cat claws as hidden items and equip when entering cat form. This
is to emulate that weapon on-hit procs cannot proc while in Cat Form,
and to be able to use as much of the baseline functionality when it
comes to weapon speed and weapon damage ranges for calculations without
having to make special cases for Cat (and Bear) Form.