Remove <source> sub-element from <proc> element.
Incorporate its data as an attribute in the <spell> element instead.

This fixes a bug where if multiple <spell> elements are defined in the
<proc> element only the last defined was propagated. Note that this only
affects Thunderfury as of this commit, effectively fixing a bug that was
introduced in 7a43cc4a456acb287b859646fbac7514b6102a21 where its 100%
nature damage proc was not propagated properly to its constructed Item.