Add a .clang-format file
This project has some code style rules written in a doc on the github
repository. It's much easier for contributors to just run clang-format
on the project instead of going to a URL and reading to make sure things
are right.

As a first step, add the .clang-format the qt-creator github repository.
I modified a few configurations to make it agree with this project:

* I added a section that implements the include sorting rules.
* I set the columnLimit to 200. This project doesn't seem to have a hard
  limit.
* I set `DerivePointerAlignment` which will try to figure out how the
  file authors put the & in, e.g., `int & a` (right|left|middle).

There are pretty widespread inconsistencies with the pointer alignment
patterns. Other than that this .clang-format seems minimally invasive
for enforcing the projects coding style.