Remove intermittently failing Shadow Bolt test.
The reason it fails is due to a circumvention of the mocked crit penalty
that is attempting to guarantee a hit, but because the Devastation
talent (prerequisite for Ruin) increases a spell-specific bonus the crit
penalty fails to reduce it (implementation detail). Fixing it is
difficult, and this test is considered to be of low value so we accept
the technical debt.