Fix restoring last toggled rotation from save file
Rotation is handled a bit weird here. We have 3 different toggleable
setups for talents, equipment and buffs. We save the rotation
information in the same fashion as the above trio, but it doens't realy
make sense to apply the rotation necessarily to any of the above three
things. So it lives separate. But adding three buttons to the GUI to
track rotations would be even clumsier than we currently have it.

So instead just save the last selected rotation to be loaded on the next
launch.

Also, this implementation is also pretty weird. The
`ClassicSimControl::set_character` method wipes the rotation database
each time it is called and rebuilds it from the rotation XML files. It
does this because the same `RotationModel` is used for each class and
thus needs to be reset while switching.

So this creates a few problems.
First, toggling from class A -> class B causes class A's rotation to be
invalid since the rotation gets `delete`ed.
Second, we don't know what the previous rotation was in the first place
in order to reselect it.

Thus, add a `QString` property for the current rotation attached to the
`Character` that will need to be updated accordingly.

A much nicer and harder future refactor would be to overhaul the
rotation system to remove this dance and have the `Character` store the
rotation database which is then loaned to the model and is thus
preserved between class switches.