Remove 'Could not find buff for' Warning (#292)
This doesn't really make sense to complain about. This warning will
show up when, for example, a rotation attempts to use an ability with a
buff but the player didn't choose that talent. This is perfectly fine
and the rotation should just move on. But this spams the log with this
error message.