Data-drive template characters instead of hardcode.
These characters are defined in Raid/template_characters.json.

This file has a similar format as the saves file, with some extensions.
The attribute "setup" for each JSON object (template character) is the
same format as found in the saves file. On top of this "setup" attribute
are minor GUI attributes "title", "class", "color".